use external types in v2 client
