+++ crates/rust-eigenda-v2-common/src/lib.rs
@@ -3,6 +3,6 @@ mod core;
 mod eigenda_cert;
 mod errors;
 
-pub use core::Payload;
+pub use core::{Blob, EncodedPayload, Payload, PayloadForm};
 pub use eigenda_cert::*;
 pub use errors::*;
